% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalysis.R
\name{plot.photofit}
\alias{plot.photofit}
\title{Plot results of a model fit}
\usage{
\method{plot}{photofit}(fit, type = "response", prob = 0.95, include_error = FALSE, ...)
}
\arguments{
\item{fit}{Object that results fof class \code{photofit}}

\item{type}{String indicated the type of figure to plot. See "details".}

\item{prob}{Probability use to calculate prediction intervals for the plot of type "response".}

\item{...}{Other arguments to the function "predict" for plots of type "response"}
}
\description{
Plot results of a model fit
}
\details{
Three types of plots are supported, depending on the string assigned to the argument
\code{type}:

1. \code{type = "response"}: This will plot the measured response curve (as points),
the median prediction and the quantile intervals for the mean and for future observations
(i.e. taking into account measurement error) for the given probability (argument \code{prob}).


2. \code{type = "marginal"}" This will be display density plots for every marginal distribution
in the posterior.

3. \code{type = "pairs"}: This will display contour plot of probability density for each
pair of parameters, the marginal distributions (as in the previous type) and the correlation
between every pair of parameters (using Pearson's correlation coefficient).
}
\examples{
data = data.frame(PAR = c(25, 50, 100, 150, 250, 350, 501, 503, 750, 1000),
                  A = c(0.51, 2.04, 4.41, 6.08, 8.03, 9.36, 9.85, 10.51, 11.14, 11.76))
fit = fitNRH(data)
summary(fit)
plot(fit)
plot(fit, type = "marginal")
plot(fit, type = "pairs")
}
