% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalysis.R
\name{predict.photofit}
\alias{predict.photofit}
\title{Sample from posterior prediction}
\usage{
\method{predict}{photofit}(
  fit,
  newdata = fit$data,
  nsamples = nrow(fit$posterior),
  include_error = FALSE
)
}
\arguments{
\item{fit}{Object of class \code{photofit}}

\item{newdata}{New data for which predictions are requested}

\item{nsamples}{Number of samples to obtain from the posterior prediction}

\item{include_error}{Indicate whether predictions should include observation error or not.
Defaults to \code{TRUE}.}
}
\value{
Matrix with sample from predictive posterior with \code{nsamples} rows.
}
\description{
Sample from posterior prediction
}
