% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAnalysis.R
\name{summary.photofit}
\alias{summary.photofit}
\title{Summary of a fitted model}
\usage{
\method{summary}{photofit}(fit, cor = NULL, include_error = TRUE)
}
\arguments{
\item{fit}{Object of class \code{photofit}}

\item{cor}{Method to calculate correlation (see \code{\link{cor}}). If
\code{NULL}, correlations will not be reported.}

\item{include_error}{Whether to include the experimental error as parameter}
}
\value{
Summary of posterior distribution
}
\description{
Summary of a fitted model
}
\details{
Compute the mean, median and typical quantiles from each posterior
marginal distribution as well as the correlation among these parameters.

The methods for calculated the correlations are \code{"pearson"},
\code{"spearman"} and \code{"kendall"}.
}
